basic renderer

+++ graph/index_buffer.go
@@ -8,17 +8,17 @@ import (
 
 type IndexBuffer struct {
 	rendererId uint32
-	count int
+	count int32
 }
 
 func NewIndexBuffer(data []uint32) *IndexBuffer {
 	ret := IndexBuffer{
-		count: len(data),
+		count: int32(len(data)),
 	}
 
 	gl.GenBuffers(1, &ret.rendererId)
 	gl.BindBuffer(gl.ELEMENT_ARRAY_BUFFER, ret.rendererId)
-	gl.BufferData(gl.ELEMENT_ARRAY_BUFFER, ret.count * int(unsafe.Sizeof(uint32(0))), gl.Ptr(data), gl.STATIC_DRAW)
+	gl.BufferData(gl.ELEMENT_ARRAY_BUFFER, int(ret.count) * int(unsafe.Sizeof(uint32(0))), gl.Ptr(data), gl.STATIC_DRAW)
 
 	return &ret
 }
@@ -35,6 +35,6 @@ func (ib *IndexBuffer) Unbind() {
 	gl.BindBuffer(gl.ELEMENT_ARRAY_BUFFER, 0)
 }
 
-func (ib *IndexBuffer) GetCount() int {
+func (ib *IndexBuffer) GetCount() int32 {
 	return ib.count
 }
